// app.client/src/api/models/CreateDocumentUrlCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateDocumentUrlCommand = {
  name?: string | null;
  description?: string | null;
  url?: string | null;
  category?: string | null;
  categoryId?: string | null;
  extension?: string | null;
  content?: string | null;
  contentType?: string | null;
  documentFileName?: string | null;
  documentType?: string | null;
  documentTypeId?: string | null;
  fileSizeBytes?: number;
  isHyperlinkAndNotFile?: boolean;
  isMarkedToDelete?: boolean;
  blobUrl?: string | null;
  documentDate?: string | null;
};