// app.client/src/api/models/UpdateDocumentUrlCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateDocumentUrlCommand = {
  documentID?: string | null;
  title?: string | null;
  description?: string | null;
  url?: string | null;
  category?: string | null;
  categoryId?: string | null;
};